envir: fix an issue when an error message is shown when the simulation in run in a read-only directory that a non-exiting file cannot be deleted.
removeFile() is silently skipping the operation if the file does not exists (checkf for ENOENT),
but when it is called on a read-only filesystem, the underlying unlink() does not return 
ENOENT, but rather a 'filesystem is read-only' error. So even if the file does not exist, it generates an error that the file cannot be deleted.

+++ src/envir/eventlogfilemgr.cc
@@ -254,9 +254,12 @@ void EventlogFileManager::close()
     fileLock = nullptr;
 }
 
+
+
 void EventlogFileManager::remove()
 {
-    removeFile(filename.c_str(), "old eventlog file");
+    if (fileExists(filename.c_str()))
+        removeFile(filename.c_str(), "old eventlog file");
     entryIndex = -1;
 }
